Move signature types to transaction crate
