+++ radix-engine-interface/src/crypto/mod.rs
@@ -1,13 +1,13 @@
-mod any;
 mod ecdsa_secp256k1;
 mod eddsa_ed25519;
 mod hash;
+mod public_key;
 mod sha2;
 mod sha3;
 
-pub use self::any::*;
 pub use self::ecdsa_secp256k1::*;
 pub use self::eddsa_ed25519::*;
 pub use self::hash::*;
+pub use self::public_key::*;
 pub use self::sha2::{sha256, sha256_twice};
 pub use self::sha3::sha3;
