+++ radix-engine-interface/src/crypto/public_key.rs
@@ -1,0 +1,28 @@
+use super::{EcdsaSecp256k1PublicKey, EddsaEd25519PublicKey};
+use crate::*;
+
+/// Represents any natively supported public key.
+#[cfg_attr(
+    feature = "serde",
+    derive(serde::Serialize, serde::Deserialize),
+    serde(tag = "type", content = "public_key")
+)]
+#[derive(
+    Debug, Clone, Copy, PartialEq, Eq, Hash, ScryptoCategorize, ScryptoEncode, ScryptoDecode,
+)]
+pub enum PublicKey {
+    EcdsaSecp256k1(EcdsaSecp256k1PublicKey),
+    EddsaEd25519(EddsaEd25519PublicKey),
+}
+
+impl From<EcdsaSecp256k1PublicKey> for PublicKey {
+    fn from(public_key: EcdsaSecp256k1PublicKey) -> Self {
+        Self::EcdsaSecp256k1(public_key)
+    }
+}
+
+impl From<EddsaEd25519PublicKey> for PublicKey {
+    fn from(public_key: EddsaEd25519PublicKey) -> Self {
+        Self::EddsaEd25519(public_key)
+    }
+}
